Remove all WrappedFunctions, replaced by macro
