+++ distage/distage-core/src/test/scala/com/github/pshirshov/izumi/distage/WrappedFunctionAnnotationTest.scala
@@ -2,7 +2,7 @@ package com.github.pshirshov.izumi.distage
 
 import com.github.pshirshov.izumi.distage.Fixtures.Case16._
 import com.github.pshirshov.izumi.distage.model.definition.Id
-import com.github.pshirshov.izumi.distage.model.functions.WrappedFunction.DIKeyWrappedFunction
+import com.github.pshirshov.izumi.distage.model.functions.DIKeyWrappedFunction
 import com.github.pshirshov.izumi.distage.model.reflection.universe.RuntimeDIUniverse._
 import org.scalatest.WordSpec
